requirements.txt
python-dotenv
streamlit
fastapi
uvicorn
chromadb
sentence-transformers
langchain
langchain-groq
langchain-community
langchain-huggingface